links_retriever/app/requirements.txt
praw
newspaper3k
pymongo
treq
twisted